# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a sample Private API",
    "version": "3.0.0",
    "title": "Node.js template for Private APIs",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "contact@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],  
  "paths": {    
    "/v1/projects": {
      "parameters": [
        {
          "$ref": "#/components/parameters/requestIdParam"
        }
      ],
      "get": {
        "summary": "Filter projects",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/offsetParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/searchParam"
          },
          {
            "$ref": "#/components/parameters/tagsParam"
          },
          {
            "$ref": "#/components/parameters/deletedsParam"
          },
          {
            "$ref": "#/components/parameters/orderParam"
          },
          {
            "$ref": "#/components/parameters/fieldsParam"
          },
          {
            "$ref": "#/components/parameters/fieldsIncludeParam"
          },
          {
            "$ref": "#/components/parameters/fieldsExcludeParam"
          }
        ],
        "tags": ["projects"]
      },
      "post": {
        "summary": "Create a new project",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": true,
                "$ref": "#/components/CreateProjectRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["projects"]
      }
    },
    "/v1/projects/{projectId}": {
      "parameters": [
        {
          "in": "path",
          "name": "projectId",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "required": true
        }
      ],
      "get": {
        "summary": "Retrieve one project by ID",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "$ref": "#/components/parameters/deletedsParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["projects"]
      },
      "put": {
        "summary": "Update a existing project",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["projects"]
      },
      "delete": {
        "summary": "Delete a existing project",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["projects"]
      }
    },    
...
    "/v1/markers": {
      "parameters": [
        {
          "$ref": "#/components/parameters/requestIdParam"
        }
      ],
      "get": {
        "summary": "Filter markers",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["markers"]
      },
      "post": {
        "summary": "Create a new marker",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Marker",
                "properties": {
                  "id": {
                    "required": false
                  },
                  "createdAt": {
                    "required": false,
                    "example": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["markers"]
      }
    },
    "/v1/markers/{markerId}": {
      "parameters": [
        {
          "in": "path",
          "name": "markerId",
          "description": "ID of marker",
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "required": true
        },
        {
          "$ref": "#/components/parameters/requestIdParam"
        }
      ],
      "get": {
        "summary": "Fetch a marker by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["markers"]
      },
      "put": {
        "summary": "Update a marker by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["markers"]
      },
      "delete": {
        "summary": "Delete a marker by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": ["markers"]
      }
    },
...
    "/v1/upload": {
      "get": {
        "summary": "Filter uploads requests",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["upload", "input"]
      },
      "post": {
        "summary": "Create a new upload request",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": ["upload", "input"]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "components": {
    "parameters": {
      "requestIdParam": {
        "in": "header",
        "name": "X-Request-Id",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "required": false,
        "description": "ID of request"
      },
...
      "offsetParam": {
        "in": "query",
        "name": "from",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "nullable": true
        },        
        "required": false,
        "description": "Records count to skip **(default=0)**"
      },
      "limitParam": {
        "in": "query",
        "name": "size",        
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "default": 20,
          "example": null
        },
        "required": false,
        "description": "Records per request **(default=20)**"
      },
      "orderParam": {
        "in": "query",
        "name": "orderBy",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "required": false,
        "nullable": true,
        "description": "Records sort order (example: \"name=asc,cronological=desc\")"
      },
      "searchParam": {
        "in": "query",
        "name": "q",
        "type": "string",
        "required": false,
        "description": "Search by record name"
      },
      "fieldsParam": {
        "in": "query",
        "name": "fields",
        "schema": {
          "type": "array",
          "items": {
            "schema": {
              "type": "string"
            }
          }
        },
        "style": "form",
        "explode": false,
        "required": false,
        "description": "Fields in response body"
      },
      "fieldsIncludeParam": {
        "in": "query",
        "name": "includeFields",
        "schema": {
          "type": "array",
          "items": {
            "schema": {
              "type": "string"
            }
          }
        },
        "style": "form",
        "explode": false,
        "required": false,
        "description": "Fields to include in response body"
      },
      "fieldsExcludeParam": {
        "in": "query",
        "name": "excludeFields",
        "schema": {
          "type": "array",
          "items": {
            "schema": {
              "type": "string"
            }
          }
        },
        "style": "form",
        "explode": false,
        "required": false,
        "description": "Fields to exclude from response body"
      },
      "tagsParam": {
        "in": "query",
        "name": "tags",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "keypair",
            "example": "name=value"
          }
        },
        "style": "simple",
        "explode": false,
        "required": false,
        "description": "URI encoded tags (example: \"name1=value1,name2=value2\")"
      },
      "deletedsParam": {
        "in": "query",
        "name": "deleteds",
        "schema": {
          "type": "boolean"
        },
        "required": false,
        "description": "Enable/display deleted records"
      }
    },
    "schemas": {
      "Marker": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "required": true,
            "description": "ID of marker"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "required": true,
            "description": "ID of project"
          },
          "name": {
            "type": "string",
            "description": "Label of marker"
          },
          "iconShape": {
            "type": "string",
            "enum": [
              "pin",
              "circle",
              "triangle",
              "square"
            ],
            "required": false,
            "default": "circle",
            "description": "Format/shape of marker icon"
          },
          "fillColor": {
            "type": "string",
            "format": "color",
            "description": "Inner color of marker icon"
          },
          "borderColor": {
            "type": "string",
            "format": "color",
            "description": "Border color of marker icon"
          },
          "isDisabled": {
            "type": "boolean",
            "required": true,
            "default": false
          },
          "tags": {
            "type": "array",
            "$ref": "#/components/Tag"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "required": false
          }
        }
      },
...
    },
...
  },
...
}